Simulator major patch 2.0

+++ melo_fwk/market_data/product.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass
 @dataclass(frozen=True)
 class Product:
 	name: str
-	block_size: int
+	block_size: float
 	cap: float
 	datastream: HLOCDataStream
